requirements.txt
discord.py==2.0.1
discord-py-interactions==4.3.3